[domain] move QuoteStripper to Domain
Use it directly in MD5Hash to strip quotes from any input.
